+++ CHANGELOG.md
@@ -12,6 +12,7 @@
  - `linetv list:station` Gets the station home (TV station) list.
  - `linetv get:station` Gets the Station Home (TV Station) data.
  - `linetv get:station --format json` Gets the Station Home (TV Station) data in JSON format.
+ - `linetv get:station --page <number>` Gets the Station Home (TV Station) data start from selected page.
 
 ### Fixed
  - Fixed `npm test` failed on Windows